--- 883-car-fleet/car-fleet.py
@@ -1,13 +1,0 @@
-class Solution:
-    def carFleet(self, target: int, position: List[int], speed: List[int]) -> int:
-        pairs = [[p, s] for p, s in zip(position, speed)]
-        stack = []
-        pairs.sort(reverse = True)
-
-        for p, s in pairs:
-            time = (target - p) / s
-            stack.append(time)
-            if len(stack) >= 2 and stack[-1] <= stack[-2]:
-                stack.pop()
-        return len(stack)
-
