+++ Array/easy/3846-minimum-operations-to-make-array-sum-divisible-by-k/README.md
