+++ Array/medium/739-daily-temperatures/daily-temperatures.py
