+++ Array/medium/74-search-a-2d-matrix/search-a-2d-matrix.py
@@ -1,31 +1,23 @@
 class Solution:
     def searchMatrix(self, matrix: List[List[int]], target: int) -> bool:
-        ROWS = len(matrix)
-        COLS = len(matrix[0])
-
-        top = 0
-        bottom = ROWS - 1
-        while top <= bottom:
-            row = (top + bottom) // 2
-            if target > matrix[row][-1]:
-                top = row + 1
-            elif target < matrix[row][0]:
-                bottom = row - 1
-            else:
-                break
-
-        if not (top <= bottom):
-            return False
-
-        row = (top + bottom) // 2
+        bottom = 0
+        top = len(matrix) - 1
         left = 0
-        right = COLS - 1
-        while left <= right:
-            m = (left + right) // 2
-            if target > matrix[row][m]:
-                left = m + 1
-            elif target < matrix[row][m]:
-                right = m - 1
+        right = len(matrix[0]) - 1
+
+        while bottom <= top:
+            midrow = (top + bottom) // 2
+            if matrix[midrow][left] <= target <= matrix[midrow][right]:
+                while left <= right:
+                    midcol = (left + right) // 2
+                    if matrix[midrow][midcol] == target:
+                        return True
+                    elif matrix[midrow][midcol] > target:
+                        right = midcol - 1
+                    else:
+                        left = midcol + 1
+            elif matrix[midrow][left] > target:
+                top = midrow - 1
             else:
-                return True
+                bottom = midrow + 1
         return False
