+++ Binary-Search/easy/792-binary-search/binary-search.py
@@ -1,6 +1,7 @@
 class Solution:
     def search(self, nums: List[int], target: int) -> int:
-        left, right = 0, len(nums) - 1
+        left = 0
+        right = len(nums) - 1
 
         while left <= right:
             mid = (left + right) // 2
