--- Matrix/medium/74-search-a-2d-matrix/Notes.md
@@ -1,1 +1,0 @@
-<h2>search-a-2d-matrix Notes</h2><hr>[ Time taken: 21 hrs 59 m 2 s ]
