--- String/easy/20-valid-parentheses/Notes.md
@@ -1,1 +1,0 @@
-<h2>valid-parentheses Notes</h2><hr>[ Time taken: 6 hrs 40 m 16 s ]
